Switch viewer default & front end logic code for adding canvases done
