+++ services/admin/src/components/shared/Switch.svelte
@@ -5,11 +5,14 @@
   export let checkVal: any;
 
   let checkStore = writable(checkVal);
+  let showDefaultStore = writable(true);
 
   setContext("checkStore", checkStore);
+  setContext("showDefaultStore", showDefaultStore);
 
   function updateMethod(checkVal: any) {
     checkStore.set(checkVal);
+    showDefaultStore.set(true);
   }
 
   $: {
