+++ services/admin/src/components/shared/SwitchCase.svelte
@@ -1,11 +1,17 @@
 <script lang="ts">
   import { getContext } from "svelte";
+  import type { Writable } from "svelte/store";
   export let caseVal: any;
   let checkStore: SvelteStore<any> = getContext("checkStore");
-  /*checkStore.subscribe((val: any) => {});*/
+  let showDefaultStore: Writable<boolean> = getContext("showDefaultStore");
+  let switchOn = false;
+  $: {
+    switchOn = $checkStore === caseVal;
+    $showDefaultStore = $showDefaultStore && !switchOn;
+  }
 </script>
 
-{#if $checkStore === caseVal}
+{#if switchOn}
   <div class="switch-viewer">
     <slot />
   </div>
